implement fast subgroup checks from B19
